+++ de.haumacher.msgbuf.generator/src/main/java/de/haumacher/msgbuf/generator/util/AbstractTypeScriptGenerator.java
@@ -1,0 +1,22 @@
+/*
+ * Copyright (c) 2026 Bernhard Haumacher et al. All Rights Reserved.
+ */
+package de.haumacher.msgbuf.generator.util;
+
+/**
+ * Base class for generators creating TypeScript code.
+ */
+public abstract class AbstractTypeScriptGenerator extends AbstractFileGenerator {
+
+	@Override
+	protected void docComment(String comment) {
+		if (comment != null && !comment.isEmpty()) {
+			line("/**");
+			for (String line : comment.split("\n")) {
+				line(" *" + (line.isEmpty() ? "" : " " + line));
+			}
+			line(" */");
+		}
+	}
+
+}